post-migration cleanup
